Use only symfony/panther when is needed
