+++ composer.json
@@ -61,6 +61,7 @@
         "psalm/plugin-symfony": "^2.0",
         "sonata-project/block-bundle": "^3.17 || ^4.0",
         "sonata-project/entity-audit-bundle": "^1.1",
+        "symfony/css-selector": "^4.4 || ^5.2",
         "symfony/panther": "^1.0",
         "symfony/phpunit-bridge": "^5.2",
         "symfony/templating": "^4.4 || ^5.2",
